[#39] 기업 공고 보기 UI 구현
보낸 제안 > 보낸 제안서 > 공고 보기 > 기업 공고
